mock-api/requirements.txt
fastapi[standard]
uvicorn[standard]
python-jose[cryptography]
pydantic-settings
